Backend/requirements.txt
django
django-cors-headers
djangorestframework
Pillow
psycopg2